refactor: enforce import best practices across all packages (closes #6)
- Use relative imports inside each package's __init__.py and modules
- Route cross-package imports through public __init__ API
- Fix circular import in embedding/builder.py using TYPE_CHECKING
- Update config.py to import EmbeddingConfig via embedding public API
