+++ src/ai_knowledge_assistant/grounding/__init__.py
@@ -1,5 +1,5 @@
-from ai_knowledge_assistant.grounding.answerability import AnswerabilityGate
-from ai_knowledge_assistant.grounding.output_validator import OutputValidator
+from .answerability import AnswerabilityGate
+from .output_validator import OutputValidator
 
 __all__ = [
     "AnswerabilityGate",
