+++ src/ai_knowledge_assistant/ingest/__init__.py
@@ -1,4 +1,4 @@
-from ai_knowledge_assistant.ingest.reader import read_files
+from .reader import read_files
 
 __all__ = [
     "read_files",
