+++ src/ai_knowledge_assistant/llm/__init__.py
@@ -1,5 +1,5 @@
-from ai_knowledge_assistant.llm.base import LLMClient, LLMConfig
-from ai_knowledge_assistant.llm.openai_client import OpenAIClient
+from .base import LLMClient, LLMConfig
+from .openai_client import OpenAIClient
 
 __all__ = [
     "LLMClient",
