+++ src/ai_knowledge_assistant/normalize/__init__.py
@@ -1,5 +1,5 @@
-from ai_knowledge_assistant.normalize.base import Chunk
-from ai_knowledge_assistant.normalize.chunker import (
+from .base import Chunk
+from .chunker import (
     get_chunks_from_files,
     split_into_chunks,
 )
