+++ src/ai_knowledge_assistant/rag/__init__.py
@@ -1,4 +1,4 @@
-from ai_knowledge_assistant.rag.prompt_builder import build_prompt
+from .prompt_builder import build_prompt
 
 __all__ = [
     "build_prompt",
