+++ src/ai_knowledge_assistant/retrieve/__init__.py
@@ -1,4 +1,4 @@
-from ai_knowledge_assistant.retrieve.retriever import Retriever
+from .retriever import Retriever
 
 __all__ = [
     "Retriever",
