+++ src/ai_knowledge_assistant/retrieve/retriever.py
@@ -2,11 +2,9 @@
 
 import logging
 
-from ai_knowledge_assistant.embedding.base import EmbeddingConfig
-from ai_knowledge_assistant.embedding.builder import EmbeddedChunk, EmbeddingBuilder
-from ai_knowledge_assistant.normalize.chunker import Chunk
-from ai_knowledge_assistant.store.faiss_store import FaissStore, ScoredChunk
-from ai_knowledge_assistant.store.vector_store import load_chunks_vectors
+from ai_knowledge_assistant.embedding import EmbeddingConfig, EmbeddedChunk, EmbeddingBuilder
+from ai_knowledge_assistant.normalize import Chunk
+from ai_knowledge_assistant.store import FaissStore, ScoredChunk, load_chunks_vectors
 
 logger = logging.getLogger(__name__)
