+++ src/ai_knowledge_assistant/store/__init__.py
@@ -1,15 +1,16 @@
-from ai_knowledge_assistant.store.base import ScoredChunk
-from ai_knowledge_assistant.store.faiss_store import FaissStore
-from ai_knowledge_assistant.store.json_store import save_chunks
-from ai_knowledge_assistant.store.vector_store import (
+from .base import ScoredChunk
+from .faiss_store import FaissStore
+from .json_store import load_chunks, save_chunks
+from .vector_store import (
     load_chunks_vectors,
     save_chunks_vectors,
 )
 
 __all__ = [
     "save_chunks",
+    "load_chunks",
     "save_chunks_vectors",
-    "ScoredChunk",
     "load_chunks_vectors",
+    "ScoredChunk",
     "FaissStore",
 ]
